answer_maker/requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==20.3.0
autopep8==1.5.6
bowler==0.9.0
cachetools==4.2.1
Cerberus==1.3.3
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
dataclasses==0.6
DateTime==4.3
decorator==5.0.7
dephell-archive==0.1.7
dephell-argparse==0.1.3
dephell-changelogs==0.0.1
dephell-discover==0.2.10
dephell-licenses==0.1.7
dephell-links==0.1.5
dephell-markers==1.0.3
dephell-pythons==0.1.15
dephell-setuptools==0.2.4
dephell-shells==0.1.5
dephell-specifier==0.2.2
dephell-venvs==0.1.18
dephell-versioning==0.1.2
docker==5.0.0
dockerpty==0.4.1
docutils==0.17
fastcore==1.3.19
fissix==20.8.0
Flask==1.1.2
Flask-Classy==0.6.10
Flask-Ext==0.1
Flask-Injector==0.12.3
future==0.18.2
google-api-core==1.26.3
google-auth==1.28.1
google-cloud-core==1.6.0
google-cloud-datastore==2.1.0
googleapis-common-protos==1.53.0
graphviz==0.16
grpcio==1.37.0
grpcio-tools==1.37.0
gunicorn==20.1.0
html5lib==1.1
idna==2.10
Inject==4.3.1
injector==0.18.4
itsdangerous==1.1.0
Jinja2==2.11.3
jp-pref==0.0.5
libcst==0.3.18
line-bot-sdk==1.19.0
m2r==0.2.1
MarkupSafe==1.1.1
mistune==0.8.4
moreorless==0.4.0
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.20.2
packaging==20.9
pandas==1.2.4
pexpect==4.8.0
pinject==0.14.1
proto-plus==1.18.1
protobuf==3.15.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-gnupg==0.4.7
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
shellingham==1.4.0
six==1.15.0
tabulate==0.8.9
toml==0.10.2
tomlkit==0.7.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
volatile==2.1.0
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==1.0.1
yapf==0.31.0
yarl==1.6.3
yaspin==1.5.0
zope.interface==5.3.0
google-cloud-secret-manager